lambda/authorizer/requirements.txt
PyJWT[crypto]==2.9.0
